requirements.txt
discord
python-dotenv
requests
pygrocy